+++ code/globalcpp/class.cpp
@@ -221,7 +221,7 @@ void *Class::operator new( size_t s )
 	if ( s == 0 )
 		return 0;
 
-	s += sizeof( unsigned int );
+	s += sizeof( size_t );
 
 #ifdef GAME_DLL
 	p = ( size_t * )gi.Malloc( s );
@@ -247,7 +247,7 @@ void *Class::operator new( size_t s )
 
 void Class::operator delete( void *ptr )
 {
-	unsigned int *p = ( ( unsigned int * )ptr  ) - 1;
+	size_t *p = ( ( size_t * )ptr  ) - 1;
 
 	totalmemallocated -= *p;
 	numclassesallocated--;
